chore: view order dto

+++ src/main/java/ro/itschool/store_management/persistence/entity/Order.java
@@ -27,7 +27,7 @@ public class Order {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private float totalPrice;
+    private double totalPrice;
     private String status;
 
     @JsonBackReference
